вынес виджеты используемые на других экрнанах
